pyproject.toml

[tool.poetry]
name = "imagecatalog"
version = "0.1.0"
description = "Create a PDF contact sheet"
authors = ["Tim Morello <tdmorello@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fpdf2 = "^2.4.6"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
flake8-black = "^0.2.3"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
flake8-pytest = "^1.3"
mypy = "^0.930"
nox = "^2021.10.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pre-commit = "^2.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
